requirements-airflow.txt
apache-airflow>=2.10.0
requests>=2.31.0
